[REFACTOR] create teaminfo.hpp and fpsent.hpp
shrinking game.hpp
