+++ inexor/fpsgame/game.hpp
@@ -15,191 +15,14 @@
 
 #include "inexor/fpsgame/gamemode.hpp"
 #include "inexor/engine/particles.hpp"
+#include "inexor/fpsgame/fpsent.hpp"
 #include "inexor/fpsgame/fpsstate.hpp"
-
+#include "inexor/fpsgame/teaminfo.hpp"
 #include "inexor/fpsgame/ai.hpp"
-
-
-/// important teamspecific declarations
-#define MAXTEAMS 128
-#define MAXNAMELEN 15  /// max player name length
-#define MAXTEAMLEN 4   /// max team name length
-#define MAXTAGLEN 8    /// max player tag length
-#define BOTTAG "Bot"   /// all bots share this tag
-
-enum
-{
-    TEAM_NONE,
-    TEAM_OWN,
-    TEAM_OPPONENT,
-    TEAM_NUM
-};
-
-/// const radar blip colors
-static const char * const teamblipcolor[TEAM_NUM] = 
-{
-    "_neutral", /// = 'gray'
-    "_blue",
-    "_red"
-};
-
-#define MAXRAYS 20
-#define EXP_SELFDAMDIV 2
-#define EXP_SELFPUSH 2.5f
-#define EXP_DISTSCALE 1.5f
-#define BOMB_DAMRAD 20
-
-// mostly players can be described with this
-struct fpsent : dynent, fpsstate
-{
-    int weight;                         // affects the effectiveness of hitpush
-    int clientnum, privilege, lastupdate, plag, ping;
-    int lifesequence;                   // sequence id for each respawn, used in damage test
-    int respawned, suicided;
-    int lastpain;
-    int lastaction, lastattackgun;
-    bool attacking;
-    int attacksound, attackchan, idlesound, idlechan;
-    int lasttaunt;
-    int lastpickup, lastpickupmillis, lastbase, lastrepammo, flagpickup, tokens;
-    vec lastcollect;
-    int frags, flags, deaths, teamkills, totaldamage, totalshots;
-    editinfo *edit;
-    float deltayaw, deltapitch, deltaroll, newyaw, newpitch, newroll;
-    int smoothmillis;
-
-    string name, tag, team, info;
-    int playermodel;
-    int fov;                            // field of view
-
-    ai::aiinfo *ai;
-    int ownernum, lastnode;
-
-    vec muzzle;
-
-    fpsent() : weight(100), clientnum(-1), privilege(PRIV_NONE), lastupdate(0), plag(0), ping(0), lifesequence(0), respawned(-1), suicided(-1), lastpain(0), attacksound(-1), attackchan(-1), idlesound(-1), idlechan(-1),
-                frags(0), flags(0), deaths(0), teamkills(0), totaldamage(0), totalshots(0), edit(NULL), smoothmillis(-1), playermodel(-1), fov(100), ai(NULL), ownernum(-1), muzzle(-1, -1, -1)
-    {
-        name[0] = team[0] = tag[0] = info[0] = 0;
-        respawn();
-    }
-    ~fpsent()
-    {
-        freeeditinfo(edit);
-        if(attackchan >= 0) inexor::sound::stopsound(attacksound, attackchan);
-        if(idlechan >= 0) inexor::sound::stopsound(idlesound, idlechan);
-        if(ai) delete ai;
-    }
-
-    /// apply push event to object's velocity vector
-    void hitpush(int damage, const vec &dir, fpsent *actor, int gun)
-    {
-        vec push(dir);
-        push.mul((actor==this && guns[gun].exprad ? EXP_SELFPUSH : 1.0f)*guns[gun].hitpush*damage/weight);
-        vel.add(push);
-    }
-
-    /// @see stopsound
-    void stopattacksound()
-    {
-        if(attackchan >= 0) inexor::sound::stopsound(attacksound, attackchan, 250);
-        attacksound = attackchan = -1;
-    }
-
-    /// @see stopsound
-    void stopidlesound()
-    {
-        if(idlechan >= 0) inexor::sound::stopsound(idlesound, idlechan, 100);
-        idlesound = idlechan = -1;
-    }
-
-    /// respawn item
-    void respawn()
-    {
-        dynent::reset();
-        fpsstate::respawn();
-        respawned = suicided = -1;
-        lastaction = 0;
-        lastattackgun = gunselect;
-        attacking = false;
-        lasttaunt = 0;
-        lastpickup = -1;
-        lastpickupmillis = 0;
-        lastbase = lastrepammo = -1;
-        flagpickup = 0;
-        tokens = 0;
-        lastcollect = vec(-1e10f, -1e10f, -1e10f);
-        stopattacksound();
-        lastnode = -1;
-    }
-};
-
-
-// team handling
-
-#define MAXNAMELEN 15
-#define MAXTEAMLEN 4
-#define MAXTEAMS 128
-
-/// some team modes allow more than 2 teams
-/// allow sorting multiple teams using team scores
-struct teamscore
-{
-    const char *team;
-    int score;
-    teamscore() {}
-    teamscore(const char *s, int n) : team(s), score(n) {}
-
-    // used for quicksort template to compare teams
-    static bool compare(const teamscore &x, const teamscore &y)
-    {
-        if(x.score > y.score) return true;
-        if(x.score < y.score) return false;
-        return strcmp(x.team, y.team) < 0;
-    }
-};
-
-/// create hashes to access hashmaps
-static inline uint hthash(const teamscore &t) { return hthash(t.team); }
-
-/// compare two teamnames
-static inline bool htcmp(const char *key, const teamscore &t) { return htcmp(key, t.team); }
-
-/// scoreboard team block description
-struct teaminfo
-{
-    char team[MAXTEAMLEN+1];
-    int frags;
-};
-
-/// create hash for hashsts
-static inline uint hthash(const teaminfo &t) { return hthash(t.team); }
-
-/// compare two team names
-static inline bool htcmp(const char *team, const teaminfo &t) { return !strcmp(team, t.team); }
 
 
 namespace game
 {
-    struct scoregroup : teamscore
-    {
-        vector<fpsent *> players;
-
-        /// Returns whether this scoregroup is a clan/playing-group, whatever.
-        /// @return the shared tag if all tags are the same, otherwise null.
-        char *sametag()
-        {
-            fpsent *prev = NULL;
-            loopv(players)
-            {
-                fpsent *p = players[i];
-                if(!p->tag[0]) return NULL; 
-                if(!prev) { prev = p; continue; }
-                if(strcmp(p->tag, prev->tag)) return NULL; //two tags are not matching
-            }
-            return prev ? prev->tag : NULL;
-        }
-    };
 
     struct clientmode
     {
